services/credit-service/requirements.txt
blinker==1.9.0
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.46.0
MarkupSafe==3.0.3
numba==0.64.0
numpy==2.4.4
packaging==26.0
pandas==3.0.1
pika==1.3.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.7
